+++ source/versionInfo.py
@@ -9,12 +9,16 @@ As there are localizable strings at module level, this can only be imported once
 To access version information for programmatic version checks before languageHandler.initialize, use the buildVersion module which contains all the non-localizable version information such as major and minor version, and version string etc.
 """
 
-from buildVersion import name, version, version_detailed
+from buildVersion import (
+	copyrightYears,
+	name,
+	url,
+	version,
+	version_detailed,
+)
 
 longName = _("NonVisual Desktop Access")
 description = _("A free and open source screen reader for Microsoft Windows")
-url = "https://www.nvaccess.org"
-copyrightYears = "2006-2026"
 copyright = _("Copyright (C) {years} NVDA Contributors").format(
 	years=copyrightYears,
 )